requirements.txt
mediapipe
opencv-python
torch
torchvision
torchaudio
ftfy
regex
tqdm
open_clip_torch
pillow